Fixed behaviour of handling internal methods/modules
